pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "A new agent project"
authors = [
    {name = "Kuanysh"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
